requirements.txt
PyJWT
requests
simpleobsws